Move Path and path param machinery to extract/path
